fix(guest): vk_id → bigint, блокировка гостя во всех гостевых флоу, LOGGING
Три фикса одного дня (23.08.2026):

1. VK ID 2.0 (2000xxxxxxxx > int4): PositiveIntegerField → PositiveBigIntegerField
   на guest.Client, discovery.DiscoveryEvent/Claim, senler.BroadcastRecipient/
   AutoBroadcastLog + миграции (public + все тенант-схемы). До этого шесть
   гостей с новыми VK ID месяцами ловили голый 500 на регистрации (594 попытки,
   найдено по логам Postgres «integer out of range»). Страховка: except DataError
   → 400 с logger.exception в ClientView.post и DiscoveryClaimView.post.

2. Блокировка гостя дырявила только парадную дверь: discovery/claim создавал
   ClientBranch заблокированному (6 живых случаев), а 16 гостевых резолвов
   не смотрели на is_active. Теперь: GuestBlocked в _create_welcome_gift и
   activate (403 в discovery-вьюхах), client__is_active=True во всех резолвах
   (fail-secure 404), get_client_profile отдаёт ту же 403, что и POST.

3. LOGGING: django.request → stderr. До этого 500-ки не логировались вообще —
   причину п.1 пришлось выкапывать из логов БД.

Тесты: +3 (блокировка в get_client_profile и welcome gift), 11 падений
BranchInfo/VKAuth — давно сломаны, падают и на чистом HEAD (проверено stash).

+++ apps/shared/discovery/models.py
@@ -28,7 +28,7 @@ class DiscoveryStage(models.TextChoices):
 class DiscoveryEvent(models.Model):
     """Событие воронки сетевого входа (open / play / claim_open)."""
 
-    vk_id = models.PositiveIntegerField(db_index=True, verbose_name='VK ID гостя')
+    vk_id = models.PositiveBigIntegerField(db_index=True, verbose_name='VK ID гостя')
     stage = models.CharField(
         max_length=16, choices=DiscoveryStage.choices, db_index=True,
         verbose_name='Стадия воронки',
@@ -61,7 +61,7 @@ class DiscoveryClaim(models.Model):
     StoryGiftEntry). redeemed_at заполняется при активации на кассе.
     """
 
-    vk_id = models.PositiveIntegerField(unique=True, verbose_name='VK ID гостя')
+    vk_id = models.PositiveBigIntegerField(unique=True, verbose_name='VK ID гостя')
     company = models.ForeignKey(
         'clients.Company',
         on_delete=models.CASCADE,
